backend/default_requirements.txt
amqp==5.2.0
asgi-runserver==0.1.3
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
black==24.8.0
boto3==1.35.10
botocore==1.35.10
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==2.5.0
Django==5.1.1
django-appconf==1.0.6
django-celery-beat==2.7.0
django-celery-email==3.0.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-cron==0.6.0
django-redis==5.4.0
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
hyperlink==21.0.0
idna==3.8
importlib_metadata==8.2.0
incremental==24.7.2
iniconfig==2.0.0
jmespath==1.0.1
jsonpickle==3.3.0
kombu==5.4.0
msgpack==1.0.8
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytest==8.3.2
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
websockets==13.0.1
yapf==0.40.2
zipp==3.20.0
zope.interface==7.0.3
